refactor: 공통 컴포넌트 위치하는 common 폴더 추가
